% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{methods}
\name{activateSubproject,Project,character-method}
\alias{activateSubproject,Project,character-method}
\title{Activate other subproject in objects of \code{\link{Project}} class}
\usage{
\S4method{activateSubproject}{Project,character}(.Object, sp, ...)
}
\arguments{
\item{.Object}{an object of class \code{\link{Project}}}

\item{sp}{character with the subproject name}
}
\description{
This method switches the between the subprojects
within the \code{\link{Project}} object
}
\details{
To check what are the subproject names
call \code{listSubprojects(config(p))}, where \code{p} is the object
of \code{\link{Project}} class
}
\examples{

}
