% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{activateSubproject}
\alias{activateSubproject}
\title{Activate other subproject in objects of \code{\link{Project-class}} class}
\usage{
activateSubproject(.Object, sp)
}
\arguments{
\item{.Object}{an object of class \code{\link{Project-class}}}

\item{sp}{character with the subproject name}
}
\description{
This method switches between the subprojects
within the \code{\link{Project-class}} object
}
\details{
To check what are the subproject names
call \code{listSubprojects(p)}, where \code{p} is the object
of \code{\link{Project-class}} class
}
\examples{
projectConfig = system.file("extdata",
"example_peps-master",
"example_subprojects1",
"project_config.yaml",
package = "pepr")
p = Project(file = projectConfig)
availSubprojects = listSubprojects(p)
activateSubproject(p,availSubprojects[1])
}
