% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadConfig.R
\name{assemblyByOrganism}
\alias{assemblyByOrganism}
\title{Mapper of organism name to genomic assembly name}
\usage{
assemblyByOrganism(config)
}
\arguments{
\item{config}{A project configuration object (parsed from a file with 
\code{yaml::yaml.load_file}).}
}
\value{
Mapping (as \code{list}) in which each name is an organism, and 
        each list element is the corresponding genomic assembly name.
}
\description{
\code{assemblyByOrganism} uses the data in a project config object to 
map organism name to genomic assembly name, supporting both the direct 
mapping within a \code{genome} section, or an encoding of this data in 
an \code{implied_columns} section.
}
\seealso{
\url{http://looper.readthedocs.io/en/latest/implied-columns.html}
}
