% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{methods}
\name{checkSection,Config-method}
\alias{checkSection,Config-method}
\title{Check for existence of a section in the Project config}
\usage{
\S4method{checkSection}{Config}(object, name)
}
\arguments{
\item{object}{object of \code{\link[pepr]{Config-class}}}

\item{name}{the name of the section to look for}
}
\value{
a logical indicating whether the section exists
}
\description{
This function checks for the section in the config YAML file. Returns \code{TRUE} if it exists or \code{FALSE} and warning otherwise.
}
