% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{methods}
\name{config,Project-method}
\alias{config,Project-method}
\title{View PEP config of the object of \code{\link{Project}} class}
\usage{
\S4method{config}{Project}(object)
}
\arguments{
\item{object}{an object of \code{\link{Project}} class}
}
\value{
a list with the config file
}
\description{
This method can be used to view the config slot of the \code{\link{Project}} class
}
