% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\title{View PEP config of the object of \code{\link{Project-class}}}
\usage{
config(object)
}
\arguments{
\item{object}{an object of \code{\link{Project-class}}}
}
\value{
a list with the config file
}
\description{
This method can be used to view the config slot of
the \code{\link{Project-class}} class
}
\examples{
projectConfig = system.file("extdata", "example_peps-cfg2",
"example_amendments1", "project_config.yaml", package="pepr")
p=Project(projectConfig)
config(p)

}
