% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{.appendAttrs}
\alias{.appendAttrs}
\title{Append constant attributes across all the samples}
\usage{
.appendAttrs(.Object)
}
\arguments{
\item{.Object}{an object of \code{\link{Project-class}}}
}
\value{
an object of \code{\link{Project-class}}
}
\description{
Append constant attributes across all the samples
}
\keyword{internal}
