% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{.applyAmendments}
\alias{.applyAmendments}
\title{Apply amendments}
\usage{
.applyAmendments(cfg, amendments = NULL)
}
\arguments{
\item{cfg}{config}

\item{amendments}{}
}
\value{

}
\description{
Overwrite and/or add Project attributes from the amendments section
}
