% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{.autoMergeDuplicatedNames}
\alias{.autoMergeDuplicatedNames}
\title{Merge samples with identical names}
\usage{
.autoMergeDuplicatedNames(.Object)
}
\arguments{
\item{.Object}{an object of \code{"\linkS4class{Project}"}}
}
\value{
an object of \code{"\linkS4class{Project}"}
}
\description{
If sample table specifies samples with non-unique names, try to merge these samples
}
\keyword{internal}
