% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.checkSection}
\alias{.checkSection}
\title{Check for a section existence in a nested list}
\usage{
.checkSection(object, sectionNames)
}
\arguments{
\item{object}{list to inspect}

\item{sectionNames}{vector or characters with sectio names to check for}
}
\value{
logical indicating whether the sections where found in the list
}
\description{
Check for a section existence in a nested list
}
\examples{
l = list(a=list(b="test"))
.checkSection(l,c("a","b"))
.checkSection(l,c("c","b"))
}
\keyword{internal}
