% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{.deriveAttrs}
\alias{.deriveAttrs}
\title{Derive attributes}
\usage{
.deriveAttrs(.Object)
}
\arguments{
\item{.Object}{an object of \code{\link{Project-class}}}
}
\value{
an object of \code{\link{Project-class}}
}
\description{
Derive attributes
}
