% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{.expandList}
\alias{.expandList}
\title{Recursively try to expand list of strings}
\usage{
.expandList(x)
}
\arguments{
\item{x}{list, possibly of strings that are paths to expand}
}
\value{
list of strings with paths expaned
}
\description{
Recursively try to expand list of strings
}
\examples{
x = list(a=list(b=list(c="~/test.txt")))
.expandList(x)
}
\keyword{internal}
