% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.expandPath}
\alias{.expandPath}
\title{Expand system path}
\usage{
.expandPath(path)
}
\arguments{
\item{path}{file path to expand. Potentially any string}
}
\value{
Expanded path or untouched string
}
\description{
This function expands system paths (the non-absolute paths become absolute)
and replaces the environment variables (e.g, \code{${HOME}})
with their values.
}
\details{
Most importantly strings that are not system paths are returned untouched
}
\examples{

string = "https://www.r-project.org/"
.expandPath(string)
path = "$HOME/my/path/string.txt"
.expandPath(path)
}
\keyword{internal}
