% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getSampleTablePathFromConfig}
\alias{.getSampleTablePathFromConfig}
\title{Get the sample table from config}
\usage{
.getSampleTablePathFromConfig(config)
}
\arguments{
\item{config}{an object of \code{"\linkS4class{Config}"}}
}
\value{
a string which specifies a path to the sample table file
}
\description{
Get the sample table from config
}
\keyword{internal}
