% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getSubSampleTablePathFromConfig}
\alias{.getSubSampleTablePathFromConfig}
\title{Get the subsample tables from config}
\usage{
.getSubSampleTablePathFromConfig(config)
}
\arguments{
\item{config}{an object of \code{"\linkS4class{Config}"}}
}
\value{
string/vector of strings/NULL depending on the configuration
}
\description{
Get the subsample tables from config
}
\keyword{internal}
