% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{.getSubscript}
\alias{.getSubscript}
\title{Get list subscript}
\usage{
.getSubscript(lst, i)
}
\arguments{
\item{lst}{list to search subscript for}

\item{i}{character or numeric to determine final list index}
}
\value{
numeric index of the requested element in the list
}
\description{
Based on available list element names and subscript value determine
index of the element requested
}
\examples{
l = list(a="a", b="b")
.getSubscript(l, 1) == .getSubscript(l, "a")
}
\keyword{internal}
