% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{.getTableIndexes}
\alias{.getTableIndexes}
\title{Set table indexes}
\usage{
.getTableIndexes(.Object, stIndex, sstIndex)
}
\arguments{
\item{.Object}{an object of \code{"\linkS4class{Project}"}}

\item{stIndex}{character string indicating a constructor-specified sample table index}

\item{sstIndex}{character string indicating a constructor-specified subsample table index}
}
\description{
Get sample and subsample table indexes and save as a slot on the Project object
}
\details{
This is the (sub)sample table index selection priority order:
\enumerate{
  \item Project constructor specified
  \item Config specified
  \item Deafult value
}
}
\keyword{internal}
