% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.handleSampleAnnotationDeprecation}
\alias{.handleSampleAnnotationDeprecation}
\title{Similarily for `sample_subannotation` `subsample_table` pair}
\usage{
.handleSampleAnnotationDeprecation(cfg)
}
\arguments{
\item{cfg}{list config}
}
\value{
list config
}
\description{
Similarily for `sample_subannotation` `subsample_table` pair
}
