% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{.hasPipIface}
\alias{.hasPipIface}
\title{Check if project defines pipeline interface}
\usage{
.hasPipIface(p)
}
\arguments{
\item{p}{object of \code{\link{Project-class}}}
}
\value{
logical indicating whether pipeline interface is defined
}
\description{
Check if project defines pipeline interface
}
