% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{.implyAttrs}
\alias{.implyAttrs}
\title{Imply attributes}
\usage{
.implyAttrs(.Object)
}
\arguments{
\item{.Object}{an object of \code{\link{Project-class}}}
}
\value{
an object of \code{\link{Project-class}}
}
\description{
Imply attributes
}
