% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{.inferProjectName}
\alias{.inferProjectName}
\title{Infer project name}
\usage{
.inferProjectName(cfg, filename)
}
\arguments{
\item{cfg}{config data}

\item{filename}{path to the config file}
}
\value{
string project name
}
\description{
Based on dedicated config section or PEP enclosing dir
}
\keyword{internal}
