% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.isAbsolute}
\alias{.isAbsolute}
\title{Determine whether a path is absolute.}
\usage{
.isAbsolute(path)
}
\arguments{
\item{path}{The path to check for seeming absolute-ness.}
}
\value{
Flag indicating whether the \code{path} appears to be absolute.
}
\description{
Determine whether a path is absolute.
}
\keyword{internal}
