% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.isCfg}
\alias{.isCfg}
\title{Config file or annotation file}
\usage{
.isCfg(filePath)
}
\arguments{
\item{filePath}{a string to examine}
}
\value{
a boolean, TRUE if indicating the path seems to be pointing to a config,
 or FALSE if the path seems to be pointing to an annotation file.
}
\description{
Determine if the input file seems to be a project
 config file (based on the file extension).
}
\keyword{internal}
