% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.isValidUrl}
\alias{.isValidUrl}
\title{Determine whether the string is a valid URL}
\usage{
.isValidUrl(str)
}
\arguments{
\item{str}{string to inspect}
}
\value{
logical indicating whether a string is a valid URL
}
\description{
Determine whether the string is a valid URL
}
\keyword{internal}
