% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.listifyDF}
\alias{.listifyDF}
\title{Listify data frame columns}
\usage{
.listifyDF(DF)
}
\arguments{
\item{DF}{an object of class data.frame}
}
\value{
an object of class data.frame
}
\description{
This function turns each data frame column into a list,
so that its cells can contain multiple elements
}
\examples{
dataFrame=mtcars
listifiedDataFrame=.listifyDF(dataFrame)
}
\keyword{internal}
