% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.listOfListToListOfDT}
\alias{.listOfListToListOfDT}
\title{Dataframify list sublists}
\usage{
.listOfListToListOfDT(list)
}
\arguments{
\item{list}{an object of class list}
}
\value{
an object of class data.frame
}
\description{
This function turns each list sublist into a data frame
}
\keyword{internal}
