% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{.loadConfig}
\alias{.loadConfig}
\title{Load the config of a PEP}
\usage{
.loadConfig(filename = NULL, amendments = NULL)
}
\arguments{
\item{filename}{file path to config file}

\item{amendments}{amendments to activate}
}
\description{
Loads a PEP config file
}
\seealso{
\url{https://pep.databio.org/}
}
\keyword{internal}
