% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{.loadSampleAnnotation}
\alias{.loadSampleAnnotation}
\title{Read sample annotation from disk}
\usage{
.loadSampleAnnotation(.Object)
}
\arguments{
\item{.Object}{an object of \code{\link{Project-class}}}
}
\value{
an object of \code{\link{Project-class}}
}
\description{
Read sample annotation from disk
}
