% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{.loadSubsampleAnnotation}
\alias{.loadSubsampleAnnotation}
\title{Load single subsample annotation}
\usage{
.loadSubsampleAnnotation(.Object, path = NULL, subsamplesTable = NULL)
}
\arguments{
\item{.Object}{an object of \code{"\linkS4class{Project}"}}

\item{path}{string, a path to the subsample table to read and incorporate}

\item{subsamplesTable}{data.table, provided from constructor if PEP is fetched from API}
}
\value{
an object of \code{"\linkS4class{Project}"}
}
\description{
Load single subsample annotation
}
\keyword{internal}
