% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.makeAbsPath}
\alias{.makeAbsPath}
\title{Create an absolute path from a primary target and a parent candidate.}
\usage{
.makeAbsPath(perhapsRelative, parent)
}
\arguments{
\item{parent}{a path to parent folder to use if target isn't absolute.}

\item{perhapsRelative:}{Path to primary target directory.}
}
\value{
Target itself if already absolute, else target nested within parent.
}
\description{
Create an absolute path from a primary target and a parent candidate.
}
