% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.matchesAndRegexes}
\alias{.matchesAndRegexes}
\title{Create a list of matched files in the system and unmatched regular expessions}
\usage{
.matchesAndRegexes(rgx)
}
\arguments{
\item{rgx}{string to expand in the system}
}
\value{
a list of all the elements after possible expansion
}
\description{
Create a list of matched files in the system and unmatched regular expessions
}
\keyword{internal}
