% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{.mergeAttrs}
\alias{.mergeAttrs}
\title{Merge samples defined in sample table with ones in subsample table(s)}
\usage{
.mergeAttrs(.Object, subsampleAannotationPaths, subsamplesTables = NULL)
}
\arguments{
\item{.Object}{an object of \code{"\linkS4class{Project}"}}

\item{subsampleAannotationPaths}{a vector of strings specifying the paths to sample}
}
\value{
an object of \code{"\linkS4class{Project}"}
}
\description{
Merge samples defined in sample table with ones in subsample table(s)
}
\keyword{internal}
