% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{.modifySamplesFetched}
\alias{.modifySamplesFetched}
\title{Perform all the sample attribute modifications for fetched PEPs}
\usage{
.modifySamplesFetched(object, subsamples)
}
\arguments{
\item{object}{an object of \code{"\linkS4class{Project}"}}
}
\value{
modified Project object
}
\description{
Perform all the sample attribute modifications for fetched PEPs
}
\keyword{internal}
