% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.printNestedList}
\alias{.printNestedList}
\title{Print a nested list}
\usage{
.printNestedList(lst, level = 0)
}
\arguments{
\item{lst}{list object to print}

\item{level}{the indentation level}
}
\value{
No return value, called for side effects
}
\description{
Prints a nested list in a way that looks nice
}
\details{
Useful for displaying the config of a PEP
}
\examples{
projectConfig = system.file("extdata",
"example_peps-master",
"example_basic",
"project_config.yaml",
package = "pepr")
p = Project(file = projectConfig)
.printNestedList(config(p),level=2)
}
\keyword{internal}
