% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{.reformat}
\alias{.reformat}
\title{Check config spec version and reformat if needed}
\usage{
.reformat(object)
}
\arguments{
\item{object}{an object of \code{\link{Config-class}}}
}
\value{
an object of \code{\link{Config-class}}
}
\description{
Check config spec version and reformat if needed
}
