% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{.removeAttrs}
\alias{.removeAttrs}
\title{Remove attributes across all the samples}
\usage{
.removeAttrs(.Object)
}
\arguments{
\item{.Object}{an object of \code{\link{Project-class}}}
}
\value{
an object of \code{\link{Project-class}}
}
\description{
Remove attributes across all the samples
}
