% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.safeFileExists}
\alias{.safeFileExists}
\title{Check whether the string is a valid URL or an existing local path}
\usage{
.safeFileExists(path)
}
\arguments{
\item{path}{string to be checked}
}
\value{
a logical indicating whether it's an existing path or valid URL
}
\description{
Check whether the string is a valid URL or an existing local path
}
\keyword{internal}
