% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.sanitizeConfig}
\alias{.sanitizeConfig}
\title{A designated place for any config transformations/sanity checks, like keys deprecation handling}
\usage{
.sanitizeConfig(cfg)
}
\arguments{
\item{cfg}{config or a section of one to sanitize}
}
\value{
sanitized config
}
\description{
A designated place for any config transformations/sanity checks, like keys deprecation handling
}
