% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetchPEP}
\alias{fetchPEP}
\title{Fetch a PEP from PEPhub using a registry path (namespace/project:tag)}
\usage{
fetchPEP(registryPath, raw = TRUE)
}
\arguments{
\item{registryPath}{a string for the PEP registry path (namespace/project:tag)}

\item{raw}{a boolean for whether to return a raw PEP}
}
\value{
a list, with sublists for config, sample_list, and subsample_list for the fetched PEP
}
\description{
Calls the PEPhub API to fetch PEPs.
}
\keyword{internal}
