% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadConfig.R
\name{fmt}
\alias{fmt}
\title{Format a string like python's format function}
\usage{
fmt(string, args)
}
\arguments{
\item{string}{String with variables encoded}

\item{args}{named list of arguments to use to populate the string}
}
\description{
Given a string with environment varibles (encoded like ${VAR}) and other
variables (encoded like {VAR}), this function will substitute both of these
and return the formatted string, like the python string format(). Other
variables are populated from a list of arguments
}
\examples{
fmt("{SRARAW}{SRR}_file", list(SRARAW = "hi", SRR="hello"))
}
