% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadConfig.R
\name{getConfigFile}
\alias{getConfigFile}
\title{used for my internal project naming scheme. 
returns a config file at a default location,
given a project name.}
\usage{
getConfigFile(project)
}
\description{
used for my internal project naming scheme. 
returns a config file at a default location,
given a project name.
}
