% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{methods}
\name{getPipelineInterface,Project-method}
\alias{getPipelineInterface,Project-method}
\title{Get the pipeline intraface}
\usage{
\S4method{getPipelineInterface}{Project}(.Object)
}
\arguments{
\item{.Object}{an object of \code{\link{Project-class}}}
}
\value{
an object of \code{\link{Config-class}}. A list-like representation of the YAML file.
}
\description{
Extracts the pipeline interface as a \code{\link{Config-class}} object
}
\examples{
projectConfig = system.file("extdata",
"example_peps-master",
"example_subprojects1",
"project_config.yaml",
package = "pepr")
p = Project(file = projectConfig)
getPipelineInterface(p)
}
