% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{getSample}
\alias{getSample}
\title{This method extracts the sample from the \code{\link{Project-class}} object}
\usage{
getSample(.Object, sampleName)
}
\arguments{
\item{.Object}{An object of Project class}

\item{sampleName}{character the name of the sample}
}
\value{
data.table one row data table with the sample associated metadata
}
\description{
This method extracts the sample from the \code{\link{Project-class}} object
}
\examples{
projectConfig = system.file(
"extdata",
"example_peps-master",
"example_basic",
"project_config.yaml",
package = "pepr"
)
p = Project(projectConfig)
sampleName = "frog_1"
getSample(p, sampleName)
}
