% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{methods}
\name{getSubsample,Project,character,character-method}
\alias{getSubsample,Project,character,character-method}
\title{This method extracts the subsample from the \code{\link{Project}} object}
\usage{
\S4method{getSubsample}{Project,character,character}(.Object, sampleName,
  subsampleName)
}
\arguments{
\item{.Object}{An object of Project class}

\item{sampleName}{character the name of the sample}

\item{subsampleName}{character the name of the subsample}
}
\value{
data.table one row data table with the subsample associated metadata
}
\description{
This method extracts the subsample from the \code{\link{Project}} object
}
