% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadConfig.R
\name{listSubprojects}
\alias{listSubprojects}
\title{Lists subprojects in a config file}
\usage{
listSubprojects(cfg)
}
\arguments{
\item{cfg}{Configuration section of a project}
}
\description{
Lists subprojects in an R list representation of a yaml config file read by
pepr
}
