% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadConfig.R
\name{load.config}
\alias{load.config}
\title{Alias for backwards compatibility}
\usage{
load.config(project = NULL, sp = NULL, filename = NULL,
  usesPathsSectio = FALSE)
}
\description{
Alias for backwards compatibility
}
