% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{makeSectionsAbsolute,Config,character,character-method}
\alias{makeSectionsAbsolute,Config,character,character-method}
\title{Make selected sections absolute using cfg path}
\usage{
\S4method{makeSectionsAbsolute}{Config,character,character}(object, sections, cfgPath)
}
\arguments{
\item{object}{Config}

\item{sections}{character set of sections to make absolute}

\item{cfgPath}{character absolute path to the config YAML file}
}
\value{
Config with selected sections made absolute
}
\description{
Make selected sections absolute using cfg path
}
