% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{makeSectionsAbsolute}
\alias{makeSectionsAbsolute}
\title{Make sections absolute in the config}
\usage{
makeSectionsAbsolute(object, sections, cfgPath)
}
\description{
Make sections absolute in the config
}
