% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{printNestedList}
\alias{printNestedList}
\title{Prints a nested list in a way that looks nice}
\usage{
printNestedList(lst, level = 0)
}
\arguments{
\item{lst}{list object to print}
}
\description{
Prints a nested list in a way that looks nice
}
