% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{Project}
\alias{Project}
\title{A class representing a Portable Encapsulated Project
This is a helper that creates the project with empty samples and config slots}
\usage{
Project(file, samples = list(), config = list())
}
\arguments{
\item{file}{project configuration yaml file}
}
\description{
A class representing a Portable Encapsulated Project
This is a helper that creates the project with empty samples and config slots
}
