% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{pullProject}
\alias{pullProject}
\title{Pull a PEP from PEPhub API}
\usage{
pullProject(registryPath = NULL)
}
\arguments{
\item{registryPath}{a string specifying a registry path from PEPhub}
}
\value{
an object of \code{"\linkS4class{Project}"}
}
\description{
This is a helper that creates the project with empty samples and config slots
}
\examples{
\dontrun{
# Requires network access to the PEPhub API
registryPath = 'databio/example:default'
p = pullProject(registryPath)
}
}
