% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{methods}
\name{samples,Project-method}
\alias{samples,Project-method}
\title{View samples in the objects of \code{\link{Project}} class}
\usage{
\S4method{samples}{Project}(object)
}
\arguments{
\item{object}{an object of \code{\link{Project}} class}
}
\value{
a data.table with the config file
}
\description{
This method can be used to view the samples slot of the \code{\link{Project}} class
}
