% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{samples}
\alias{samples}
\title{View samples in the objects of \code{\link{Project-class}} class}
\usage{
samples(object)
}
\arguments{
\item{object}{an object of \code{\link{Project-class}} class}
}
\value{
a data.table with the with metadata about samples
}
\description{
This method can be used to view the samples slot
of the \code{\link{Project-class}} class
}
\examples{
projectConfig = system.file("extdata", "example_peps-master",
"example_subprojects1", "project_config.yaml", package="pepr")
p=Project(projectConfig)
samples(p)

}
