% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{sampleTable}
\alias{sampleTable}
\title{View samples in the objects of \code{\link{Project-class}}}
\usage{
sampleTable(object)
}
\arguments{
\item{object}{an object of \code{\link{Project-class}}}
}
\value{
a data.table with the with metadata about samples
}
\description{
This method can be used to view the samples slot
of the \code{\link{Project-class}} class
}
\examples{
projectConfig = system.file("extdata", "example_peps-master",
"example_amendments1", "project_config.yaml", package="pepr")
p=Project(projectConfig)
sampleTable(p)

}
