% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{saveJWT}
\alias{saveJWT}
\title{Save or update JWT}
\usage{
saveJWT(jwt)
}
\arguments{
\item{jwt}{a string for the new jwt to save}
}
\value{
a boolean for whether the authentication token was saved
}
\description{
Save or update local authentication token to fetch private PEPs
}
