% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{saveProject}
\alias{saveProject}
\title{Save a modified PEP Project to a local directory}
\usage{
saveProject(project = NULL, outputDir = getwd(), overwrite = FALSE)
}
\arguments{
\item{project}{a PEP Project}

\item{overwrite}{a boolean for whether to overwrite an existing project at the output directory}

\item{projectDir}{a string for the output directory, defaults to current working directory}

\item{sampleTableIndex}{a string indicating the sample attribute that is used}
}
\value{
a boolean, TRUE if the save was successful and FALSE if otherwise
}
\description{
This is a helper that saves a PEP Project to a local output directory
}
