% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{select-config}
\alias{select-config}
\alias{[,Config-method}
\alias{[[,Config-method}
\alias{$,Config-method}
\title{Access \code{"\linkS4class{Config}"} object elements}
\usage{
\S4method{[}{Config}(x, i)

\S4method{[[}{Config}(x, i)

\S4method{$}{Config}(x, name)
}
\arguments{
\item{x}{a \code{"\linkS4class{Config}"} object.}

\item{i}{position of the identifier or the name of the identifier itself.}

\item{name}{name of the element to access.}
}
\value{
An element held in \code{"\linkS4class{Config}"} object
}
\description{
You can subset \linkS4class{Config} by identifier or by position using the
\code{`[`}, \code{`[[`} or \code{`$`} operator.
The string will be expanded if it's a path.
}
\examples{
projectConfig = system.file("extdata", "example_peps-master",
"example_amendments1", "project_config.yaml", package="pepr")
c=Config(projectConfig)
c[[2]]
c[2]
c[["sample_table"]]
c$sample_table

}
