% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{methods}
\name{checkSection,Config-method}
\alias{checkSection,Config-method}
\title{Check for existence of a section in the Project config}
\usage{
\S4method{checkSection}{Config}(object, sectionNames)
}
\arguments{
\item{object}{object of \code{\link[pepr]{Config-class}}}

\item{sectionNames}{the name of the section or names of the nested sections to look for}
}
\value{
a logical indicating whether the section exists
}
\description{
This function checks for the section/nested sections in the config YAML file. Returns \code{TRUE} if it exist(s) or \code{FALSE} otherwise.
}
\examples{
projectConfig = system.file("extdata", "example_peps-master",
"example_subprojects1", "project_config.yaml", package="pepr")
p=Project(projectConfig)
checkSection(config(p),sectionNames = c("subprojects","newLib","metadata"))
}
