% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{config}
\alias{config}
\alias{config,Project-method}
\title{Extract \code{"\linkS4class{Project}"}}
\usage{
config(object)

\S4method{config}{Project}(object)
}
\arguments{
\item{object}{an object of \code{"\linkS4class{Project}"}}
}
\value{
project config
}
\description{
This method can be used to view the config slot of
the \code{"\linkS4class{Project}"} class
}
\section{Methods (by class)}{
\itemize{
\item \code{config(Project)}: Extract \code{"\linkS4class{Project}"} of the object of \code{"\linkS4class{Project}"}

}}
\examples{
projectConfig = system.file("extdata", "example_peps-master",
"example_amendments1", "project_config.yaml", package="pepr")
p=Project(projectConfig)
config(p)

}
