% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{.deriveAttrs}
\alias{.deriveAttrs}
\title{Derive attributes}
\usage{
.deriveAttrs(.Object)
}
\arguments{
\item{.Object}{an object of \code{"\linkS4class{Project}"}}
}
\value{
an object of \code{"\linkS4class{Project}"}
}
\description{
Derive attributes
}
\keyword{internal}
