% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{.duplicateAttrs}
\alias{.duplicateAttrs}
\title{Duplicate a selected attribute across all the samples}
\usage{
.duplicateAttrs(.Object)
}
\arguments{
\item{.Object}{an object of \code{"\linkS4class{Project}"}}
}
\value{
an object of \code{"\linkS4class{Project}"}
}
\description{
Duplicate a selected attribute across all the samples
}
\keyword{internal}
