% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{.implyAttrs}
\alias{.implyAttrs}
\title{Imply attributes}
\usage{
.implyAttrs(.Object)
}
\arguments{
\item{.Object}{an object of \code{"\linkS4class{Project}"}}
}
\value{
an object of \code{"\linkS4class{Project}"}
}
\description{
Imply attributes
}
\keyword{internal}
