% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{.loadSampleAnnotation}
\alias{.loadSampleAnnotation}
\title{Read sample table from disk}
\usage{
.loadSampleAnnotation(sampleTablePath)
}
\arguments{
\item{sampleTablePath}{a character string indicating a path to the sample table}
}
\value{
an data.frame with samples; one sample per row
}
\description{
Read sample table from disk
}
\keyword{internal}
