% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{.loadSampleAnnotation}
\alias{.loadSampleAnnotation}
\title{Read sample annotation from disk}
\usage{
.loadSampleAnnotation(.Object)
}
\arguments{
\item{.Object}{an object of \code{"\linkS4class{Project}"}}
}
\value{
an object of \code{"\linkS4class{Project}"}
}
\description{
Read sample annotation from disk
}
