% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{.mergeAttrs}
\alias{.mergeAttrs}
\title{Merge samples defined in sample table with ones in subsample table}
\usage{
.mergeAttrs(.Object)
}
\arguments{
\item{.Object}{an object of \code{\link{Project-class}}}
}
\value{
an object of \code{\link{Project-class}}
}
\description{
Merge samples defined in sample table with ones in subsample table
}
