% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{.mergeAttrs}
\alias{.mergeAttrs}
\title{Merge samples defined in sample table with ones in subsample table(s)}
\usage{
.mergeAttrs(.Object)
}
\arguments{
\item{.Object}{an object of \code{"\linkS4class{Project}"}}
}
\value{
an object of \code{"\linkS4class{Project}"}
}
\description{
Merge samples defined in sample table with ones in subsample table(s)
}
