% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{.modifySamples}
\alias{.modifySamples}
\title{Perform all the sample addribute modifications}
\usage{
.modifySamples(object)
}
\arguments{
\item{object}{an object of \code{"\linkS4class{Project}"}}
}
\value{
modified Project object
}
\description{
Perform all the sample addribute modifications
}
