% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{.reformat}
\alias{.reformat}
\title{Check config spec version and reformat if needed}
\usage{
.reformat(object)
}
\arguments{
\item{object}{an object of \code{"\linkS4class{Config}"}}
}
\value{
an object of \code{"\linkS4class{Config}"}
}
\description{
Check config spec version and reformat if needed
}
\keyword{internal}
