% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{.removeAttrs}
\alias{.removeAttrs}
\title{Remove attributes across all the samples}
\usage{
.removeAttrs(.Object)
}
\arguments{
\item{.Object}{an object of \code{"\linkS4class{Project}"}}
}
\value{
an object of \code{"\linkS4class{Project}"}
}
\description{
Remove attributes across all the samples
}
\keyword{internal}
