% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{methods}
\name{getSubsample,Project,character,character-method}
\alias{getSubsample,Project,character,character-method}
\title{This method extracts the subsample from the \code{\link{Project-class}} object}
\usage{
\S4method{getSubsample}{Project,character,character}(.Object, sampleName,
  subsampleName)
}
\arguments{
\item{.Object}{An object of Project class}

\item{sampleName}{character the name of the sample}

\item{subsampleName}{character the name of the subsample}
}
\value{
data.table one row data table with the subsample associated metadata
}
\description{
This method extracts the subsample from the \code{\link{Project-class}} object
}
\examples{
projectConfig = system.file(
"extdata",
"example_peps-master",
"example_subannotation1",
"project_config.yaml",
package = "pepr"
)
p = Project(projectConfig)
sampleName = "frog_1"
subsampleName = "sub_a"
getSubsample(p, sampleName, subsampleName)
}
