% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{listAmendments}
\alias{listAmendments}
\alias{listAmendments,Project-method}
\title{List amendments}
\usage{
listAmendments(.Object)

\S4method{listAmendments}{Project}(.Object)
}
\arguments{
\item{.Object}{an object of \code{"\linkS4class{Project}"}}
}
\value{
names of the available amendments
}
\description{
Lists available amendments within a \code{"\linkS4class{Project}"} object.
}
\details{
The amendments can be activated by passing their names to the  \code{\link{activateAmendments}} method
}
\section{Methods (by class)}{
\itemize{
\item \code{listAmendments(Project)}: list amendments in a \code{"\linkS4class{Project}"} object

}}
\examples{
projectConfig = system.file("extdata",
"example_peps-master",
"example_amendments1",
"project_config.yaml",
package = "pepr")
p = Project(file = projectConfig)
availAmendemtns = listAmendments(p)
}
