% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{listAmendments}
\alias{listAmendments}
\title{Lists amendments in a \code{\link{Project-class}} object}
\usage{
listAmendments(.Object)
}
\arguments{
\item{.Object}{an object of \code{\link{Project-class}}}
}
\value{
names of the available amendments
}
\description{
Lists available amendments within a \code{\link{Project-class}} object.
}
\details{
The amendments can be activated by passing their names to the  \code{\link{activateAmendments}} method
}
\examples{
projectConfig = system.file("extdata",
"example_peps-master",
"example_amendments1",
"project_config.yaml",
package = "pepr")
p = Project(file = projectConfig)
availAmendemtns = listAmendments(p)
activateAmendments(p,availAmendemtns[1])
}
