% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadConfig.R
\name{listSubprojects}
\alias{listSubprojects}
\title{Lists subprojects in a Projec object}
\usage{
listSubprojects(project)
}
\arguments{
\item{project}{an object of \code{\link{Project-class}} class}
}
\value{
names of the available subprojects
}
\description{
Lists avialable subprojects within a \code{\link{Project-class}} object.
}
\details{
The subprojects can be activated by passing their names 
to the \code{\link{Project-class}} object constructor (\code{\link{Project-class}})
}
