% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{listSubprojects}
\alias{listSubprojects}
\title{Lists subprojects in a \code{\link{Project-class}} object}
\usage{
listSubprojects(.Object)
}
\arguments{
\item{.Object}{an object of \code{\link{Project-class}} class}
}
\value{
names of the available subprojects
}
\description{
Lists available subprojects within a \code{\link{Project-class}} object.
}
\details{
The subprojects can be activated by passing their names
to the \code{\link{Project-class}} object constructor (\code{\link{Project}})
}
