% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadConfig.R
\name{loadConfig}
\alias{loadConfig}
\title{Load the config of a PEP}
\usage{
loadConfig(filename = NULL, sp = NULL)
}
\arguments{
\item{filename}{file path to config file}

\item{sp}{Subproject to activate}
}
\description{
Loads a \code{project_config.yaml} file
}
