% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{makeSectionsAbsolute}
\alias{makeSectionsAbsolute}
\alias{makeSectionsAbsolute,Config,character,character-method}
\title{Make selected sections absolute using config path}
\usage{
makeSectionsAbsolute(object, sections, cfgPath)

\S4method{makeSectionsAbsolute}{Config,character,character}(object, sections, cfgPath)
}
\arguments{
\item{object}{\code{"\linkS4class{Config}"}}

\item{sections}{character set of sections to make absolute}

\item{cfgPath}{character absolute path to the config YAML file}
}
\value{
Config with selected sections made absolute
}
\description{
Make selected sections absolute using config path
}
\section{Methods (by class)}{
\itemize{
\item \code{makeSectionsAbsolute(
  object = Config,
  sections = character,
  cfgPath = character
)}: Make selected sections absolute using config path from \code{"\linkS4class{Project}"}

}}
