% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\docType{class}
\name{Project-class}
\alias{Project-class}
\title{Portable Encapsulated Project object}
\description{
Provides an in-memory representation and functions to access project
configuration and sample annotation values for a PEP.
}
\details{
Can be created with the constructor: \code{"\linkS4class{Project}"}
}
\section{Slots}{

\describe{
\item{\code{file}}{character vector path to config file on disk.}

\item{\code{samples}}{a data table object holding the sample metadata}

\item{\code{config}}{a list object holding contents of the config file}

\item{\code{sampleNameAttr}}{a string indicating the sample attribute that is used
to index the sample table}

\item{\code{subSampleNameAttr}}{a string indicating the sample attribute that is used
to index the sample table}
}}

