% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{class}
\name{Project-class}
\alias{Project-class}
\title{Portable Encapsulated Project object}
\description{
Provides an in-memory representation and functions to access project
configuration and sample annotation values for a PEP.
}
\section{Slots}{

\describe{
\item{\code{file}}{A logical keeping track of something.}

\item{\code{samples}}{An integer specifying something else.}

\item{\code{config}}{A data.frame holding some data.}
}}

