% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{Project}
\alias{Project}
\title{The constructor of a class representing a Portable Encapsulated Project}
\usage{
Project(
  file = NULL,
  amendments = NULL,
  sampleTableIndex = NULL,
  subSampleTableIndex = NULL
)
}
\arguments{
\item{file}{a string specifying a path to a project configuration YAML file}

\item{amendments}{a string with the amendments names to be activated}

\item{sampleTableIndex}{a string indicating the sample attribute that is used
to index the sample table}

\item{subSampleTableIndex}{a string indicating the sample attribute that is used
to index the sample table}
}
\value{
an object of \code{"\linkS4class{Project}"}
}
\description{
This is a helper that creates the project with empty samples and config slots
}
\examples{
projectConfig = system.file("extdata", "example_peps-master",
"example_amendments1", "project_config.yaml", package="pepr")
p=Project(projectConfig)
}
