% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{Project}
\alias{Project}
\title{The constructor of a class representing a Portable Encapsulated Project}
\usage{
Project(file = character(), subproject = character())
}
\arguments{
\item{file}{a character with project configuration yaml file}

\item{samples}{a data table object holding the sample metadata}

\item{config}{a list object holding contents of the config file}
}
\description{
This is a helper that creates the project with empty samples and config slots
}
