% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{Project}
\alias{Project}
\title{The constructor of a class representing a Portable Encapsulated Project}
\usage{
Project(file = character(), subproject = character())
}
\arguments{
\item{file}{a character with project configuration yaml file}

\item{subproject}{a character with the subproject name to be activated}
}
\description{
This is a helper that creates the project with empty samples and config slots
}
\examples{
projectConfig = system.file("extdata", "example_peps-master",
"example_subprojects1", "project_config.yaml", package="pepr")
p=Project(projectConfig)
}
