% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{Project}
\alias{Project}
\title{The constructor of a class representing a Portable Encapsulated Project}
\usage{
Project(file = NULL, sampleTable = NULL, amendments = NULL)
}
\arguments{
\item{file}{a character specifying a path to a project configuration YAML file}

\item{sampleTable}{a character specifying a path to the sample table path. It's disregarded if the `file` argument is provided.}

\item{amendments}{a character with the amendments names to be activated}
}
\description{
This is a helper that creates the project with empty samples and config slots
}
\examples{
projectConfig = system.file("extdata", "example_peps-master",
"example_amendments1", "project_config.yaml", package="pepr")
p=Project(projectConfig)
}
