% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{Project}
\alias{Project}
\title{The constructor of a class representing a Portable Encapsulated Project}
\usage{
Project(
  file = NULL,
  sampleTable = NULL,
  subSampleTables = NULL,
  amendments = NULL
)
}
\arguments{
\item{file}{a string specifying a path to a project configuration YAML file}

\item{sampleTable}{a string specifying a path to the sample table. It's disregarded if `file` argument is provided.}

\item{subSampleTables}{a vector of strings specifying paths to the subsample tables. It's disregarded if `file` argument is provided.}

\item{amendments}{a string with the amendments names to be activated}
}
\description{
This is a helper that creates the project with empty samples and config slots
}
\examples{
projectConfig = system.file("extdata", "example_peps-master",
"example_amendments1", "project_config.yaml", package="pepr")
p=Project(projectConfig)
}
