% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{Project}
\alias{Project}
\title{The constructor of a class representing a Portable Encapsulated Project}
\usage{
Project(file = NULL, amendments = NULL)
}
\arguments{
\item{file}{a character with project configuration yaml file}

\item{amendments}{a character with the amendments names to be activated}
}
\description{
This is a helper that creates the project with empty samples and config slots
}
\examples{
projectConfig = system.file("extdata", "example_peps-master",
"example_amendments1", "project_config.yaml", package="pepr")
p=Project(projectConfig)
}
